chore(alerts): extract detector logic from trends.py into detector.py
Split the 926-line trends.py into two focused modules:
- trends.py (~490 lines): threshold-based alert checking and helpers
- detector.py (~450 lines): anomaly detection checking, simulation,
  and detector configuration constants

Shared helpers (TrendResult, _has_breakdown, _drop_incomplete_current_interval,
_pick_series_result, _is_non_time_series_trend) remain in trends.py.
